feat: improve timestamp handling and add new utilities
- Enhanced `Timestamps` class with several new utility methods like `__contains__`,
  `get_indexes`, and `get_nearest_stamp` for better timestamp operations.
- Introduced `_check_precondition_nanoseconds_are_integers` to ensure input types
  are validated for timestamp-related calculations.
- Updated and added robust docstrings for improved clarity and developer guidance.

Issue TCT-52

+++ tests/tests_temporal_tools/conftest.py
@@ -21,7 +21,7 @@ def setup_mock_timestamps() -> np.ndarray:
         1711038330436485488,
     ]
 
-    return np.array(mock_timestamps)
+    return np.array(mock_timestamps, dtype=int)
 
 
 @pytest.fixture
@@ -48,7 +48,7 @@ def setup_mock_ros_timestamps():
 
     mock_container = {
         "feature_name": "/mock_topic",
-        "timestamps": np.array(mock_timestamps),
+        "timestamps": np.array(mock_timestamps, dtype=int),
     }
 
     return mock_container
